Increase max number of TAIs/eNB from 16 to 256 (#3)
This change allows us to scale up to one /24 subnet's worth of KeyLTE and associated RAN. Note that this limit applies not just to actively-connected UPS/RAN but the config itself. When we need to scale beyond this, we can always increase these numbers further, or (preferable) spin up a new CPS and scale horizontally. Note that this change does increase some data-structure size for a total increase in active memory use of 2.5MB.

+++ lib/pfcp/context.c
@@ -403,7 +403,7 @@ int ogs_pfcp_context_parse_config(const char *local, const char *remote)
                         const char *hostname[OGS_MAX_NUM_OF_HOSTNAME];
                         uint16_t port = self.pfcp_port;
                         uint16_t tac[OGS_MAX_NUM_OF_TAI] = {0,};
-                        uint8_t num_of_tac = 0;
+                        uint16_t num_of_tac = 0;
                         const char *dnn[OGS_MAX_NUM_OF_DNN];
                         uint8_t num_of_dnn = 0;
                         uint32_t e_cell_id[OGS_MAX_NUM_OF_CELL_ID] = {0,};
